Updated folder structure and added some comments
Added comments to ClampRotate.cs
