+++ Assets/Scripts/UI.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: cbf307c79a8b44647a0e724e8f4c5d12
+folderAsset: yes
+timeCreated: 1523882775
+licenseType: Free
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
