+++ Assets/Scripts/UI/UINotification.cs
@@ -4,7 +4,7 @@ using TOH.Core;
 using UnityEngine;
 using UnityEngine.UI;
 
-public class Notification : MonoBehaviour 
+public class UINotification : MonoBehaviour 
 {
 	[SerializeField]
 	private Text text;
